+++ src/dictionarly/helpers/get-searched-word-content.helper.ts
@@ -1,0 +1,10 @@
+import { isCharOK } from './is-char-ok.helper'
+
+export const getSearchedWordContent = (
+  char: string,
+  index: number,
+  searchedWord: string,
+  word: string,
+) => {
+  return `${char}${index === searchedWord.length - 1 && isCharOK(searchedWord, word)(index) ?  '...' : ''}`
+}