+++ src/dictionarly/helpers/index.ts
@@ -1,3 +1,4 @@
 export { appendSortedWords } from './append-sorted-words.helper'
 export { getDictionaryWords } from './get-dictionary-words.helper'
 export { isCharOK } from './is-char-ok.helper'
+export { getSearchedWordContent } from './get-searched-word-content.helper'